# Conflicts:
#   controllers/appointment.go

+++ controllers/appointment.go
@@ -21,11 +21,15 @@ func InitialiseAppointment(client mqtt.Client) {
 	tokenCancel := client.Subscribe("grp20/req/appointment/delete", byte(0), func(c mqtt.Client, m mqtt.Message) {
 
 		var payload schemas.Appointment
-		err := json.Unmarshal(m.Payload(), &payload)
-		if err != nil {
-			panic(err)
+        var returnData Res
+		err1 := json.Unmarshal(m.Payload(), &payload)
+		err2 := json.Unmarshal(m.Payload(), &returnData)
+		if (err1 != nil) && (err2 != nil) {
+			//send mqtt message 400 bad request
+			fmt.Printf("400 - Bad request")
+		} else {
+			go CancelAppointment(payload.ID, returnData, client)
 		}
-		go CancelAppointment(payload.ID, client)
 	})
 	if tokenCancel.Error() != nil {
 		panic(tokenCancel.Error())
@@ -34,12 +38,15 @@ func InitialiseAppointment(client mqtt.Client) {
 	tokenCreate := client.Subscribe("grp20/req/appointment/create", byte(0), func(c mqtt.Client, m mqtt.Message) {
 
 		var payload schemas.Appointment
-		err := json.Unmarshal(m.Payload(), &payload)
-		if err != nil {
-			panic(err)
+        var returnData Res
+		err1 := json.Unmarshal(m.Payload(), &payload)
+		err2 := json.Unmarshal(m.Payload(), &returnData)
+		if (err1 != nil) && (err2 != nil) {
+			//send mqtt message 400 bad request
+			fmt.Printf("400 - Bad request")
+		} else {
+			go CreateAppointment(payload, returnData, client)
 		}
-		go CreateAppointment(payload, client)
-
 	})
 	if tokenCreate.Error() != nil {
 		panic(tokenCreate.Error())
@@ -48,11 +55,15 @@ func InitialiseAppointment(client mqtt.Client) {
 	tokenGetAllForUser := client.Subscribe("grp20/req/appointment/get", byte(0), func(c mqtt.Client, m mqtt.Message) {
 
 		var payload schemas.Appointment
-		err := json.Unmarshal(m.Payload(), &payload)
-		if err != nil {
-			panic(err)
+		var returnData Res
+		err1 := json.Unmarshal(m.Payload(), &payload)
+		err2 := json.Unmarshal(m.Payload(), &returnData)
+		if (err1 != nil) && (err2 != nil) {
+			//send mqtt message 400 bad request
+			fmt.Printf("400 - Bad request")
+		} else {
+			go GetAllForUser(payload.Patient_id, returnData, client)
 		}
-		go GetAllForUser(payload.Patient_id, client)
 	})
 	if tokenGetAllForUser.Error() != nil {
 		panic(tokenGetAllForUser.Error())
@@ -61,11 +72,16 @@ func InitialiseAppointment(client mqtt.Client) {
 	tokenDelete := client.Subscribe("appointmentservice/internal/delete", byte(0), func(c mqtt.Client, m mqtt.Message) {
 
 		var payload schemas.Appointment
-		err := json.Unmarshal(m.Payload(), &payload)
-		if err != nil {
-			panic(err)
+        var returnData Res
+
+		err1 := json.Unmarshal(m.Payload(), &payload)
+		err2 := json.Unmarshal(m.Payload(), &returnData)
+		if (err1 != nil) && (err2 != nil) {
+			//send mqtt message 400 bad request or ignore due to internal?
+			fmt.Printf("400 - Bad request")
+		} else {
+			go DeleteAppointment(payload.ID, returnData, client)
 		}
-		go DeleteAppointment(payload.ID, client)
 	})
 	if tokenDelete.Error() != nil {
 		panic(tokenDelete.Error())
@@ -73,9 +89,7 @@ func InitialiseAppointment(client mqtt.Client) {
 
 }
 
-func DeleteAppointment(id primitive.ObjectID, client mqtt.Client) bool {
-	var message string
-	var code string
+func DeleteAppointment(id primitive.ObjectID, returnData Res, client mqtt.Client) bool {
 	var returnVal bool
 
 	col := getAppointmentCollection()
@@ -87,25 +101,22 @@ func DeleteAppointment(id primitive.ObjectID, client mqtt.Client) bool {
 
 	if result.DeletedCount == 1 {
 
-		message = `{"message": "User deleted"`
-		code = "200"
+        returnData.Message = "Appointment deleted"
+        returnData.Status = 200
 
-		message = AddCodeStringJson(message, code)
 	} else {
-		message = `{"message": "User not found"`
-		code = "404"
-		message = AddCodeStringJson(message, code)
+        returnData.Message = "Appointment not found"
+        returnData.Status = 404
 	}
 
-	client.Publish("grp20/res/dentist/delete", byte(0), false, message)
+    PublishReturnMessage(returnData, "grp20/res/dentist/delete", client)
+
 	returnVal = true
 	return returnVal
 }
 
-func GetAllForUser(id primitive.ObjectID, client mqtt.Client) bool {
+func GetAllForUser(id primitive.ObjectID, returnData Res,client mqtt.Client) bool {
 
-	var message string
-	var code string
 	var returnVal bool
 
 	col := getAppointmentCollection()
@@ -114,8 +125,8 @@ func GetAllForUser(id primitive.ObjectID, client mqtt.Client) bool {
 	cursor, err := col.Find(context.TODO(), filter)
 
 	if err != nil {
-		message = `{"message": "Error"`
-		code = "500"
+        returnData.Message = "Error"
+        returnData.Status = 500
 	}
 
 	defer cursor.Close(context.TODO())
@@ -126,61 +137,70 @@ func GetAllForUser(id primitive.ObjectID, client mqtt.Client) bool {
 		var appointment schemas.Appointment
 
 		if err := cursor.Decode(&appointment); err != nil {
-			message = `{"message": "Error"`
-			code = "500"
+            returnData.Message = "Error"
+            returnData.Status = 500
 			panic(err)
 		}
 
 		appointments = append(appointments, appointment)
 	}
-	jsonData, err := json.Marshal(appointments)
-	fmt.Printf(string(jsonData))
 	returnVal = true
 
-	code = "200"
-	message = AddCodeStringJson(string(jsonData), code)
+    returnData.Status = 200
+    returnData.Appointments = &appointments
 
-	client.Publish("grp20/res/appointment/get", 0, false, message)
+    PublishReturnMessage(returnData, "grp20/res/appointment/get", client)
+
 
 	return returnVal
 }
 
-func CreateAppointment(payload schemas.Appointment, client mqtt.Client) bool {
-	var message string
-	var code string
-	var returnVal bool
+func CreateAppointment(payload schemas.Appointment, returnData Res, client mqtt.Client) bool {
+
+	if appointmentExist(payload) {
+		//send conflict - http status code
+		return false
+	}
+
+	if payload.Start_time > payload.End_time {
+		//send mqtt message, not valid format for start and end, should be start<end
+		return false
+	}
 
 	col := getAppointmentCollection()
 
 	result, err := col.InsertOne(context.TODO(), payload)
+    payload.ID = result.InsertedID.(primitive.ObjectID)
 	_ = result
 
 	if err != nil {
-		code = "500"
-		message = `{"message": "Appointmend could not be created"`
-		client.Publish("grp20/res/appointment/create", 0, false, message)
-		panic(err)
+        returnData.Status = 500
+        returnData.Message = "Appointment could not be created"
+        PublishReturnMessage(returnData, "grp20/res/appointment/create", client)
+		return false
 	}
 
-	message = `{"message": "Appointment booked"`
-	code = "200"
-	returnVal = true
+    returnData.Message = "Appointment booked"
+    returnData.Status = 200
+    returnData.Appointment = &payload
 
-	message = AddCodeStringJson(message, code)
-
-	client.Publish("grp20/res/appointment/create", 0, false, message)
-	return returnVal
+    PublishReturnMessage(returnData, "grp20/res/appointment/create", client)
+	return true
 }
 
-func CancelAppointment(id primitive.ObjectID, client mqtt.Client) bool {
-	var message string
-	var code string
+func CancelAppointment(id primitive.ObjectID, returnData Res, client mqtt.Client) bool {
 	var returnVal bool
 	appointment := &schemas.Appointment{}
 
 	col := getAppointmentCollection()
 	filter := bson.M{"_id": id}
 	data := col.FindOne(context.TODO(), filter)
+
+	if data.Err() == mongo.ErrNoDocuments {
+		//send 404 message
+		return false
+	}
+
 	result, err := col.DeleteOne(context.TODO(), filter)
 
 	if err != nil {
@@ -189,7 +209,7 @@ func CancelAppointment(id primitive.ObjectID, client mqtt.Client) bool {
 
 	if result.DeletedCount == 1 {
 
-		code = "200"
+        returnData.Status = 200
 
 		data.Decode(appointment)
 
@@ -205,26 +225,42 @@ func CancelAppointment(id primitive.ObjectID, client mqtt.Client) bool {
 			panic(err)
 		}
 
-		message = string(jsonData)
+        message := string(jsonData)
 
 		client.Publish("appointmentservice/internal/migrate", 0, false, message)
 
-		message = `{"message": "Appointment Canceled"`
-		message = AddCodeStringJson(message, code)
+        returnData.Message = "Appointment Canceled"
 
 		returnVal = true
 
 	} else {
 
-		code = "404"
-		message = `{"message": "Appointment not found"`
-		message = AddCodeStringJson(message, code)
+        returnData.Status = 404
+        returnData.Message = "Appointment not found"
 
 		returnVal = false
 	}
 
-	client.Publish("grp20/res/appointment/delete", 0, false, message)
+    PublishReturnMessage(returnData, "grp20/res/appointment/delete", client)
 	return returnVal
+}
+
+func appointmentExist(payload schemas.Appointment) bool {
+	col := getAppointmentCollection()
+
+	filter := bson.M{
+		"dentist_id": payload.Dentist_id,
+		"patient_id": payload.Patient_id,
+		"start_time": payload.Start_time,
+		"end_time":   payload.End_time,
+	}
+
+	count, err := col.CountDocuments(context.Background(), filter)
+	if err != nil {
+		return true
+	}
+
+	return count > 0
 }
 
 func getAppointmentCollection() *mongo.Collection {
